·“翔鹤改二→翔鹤改二甲”中断检查修改为“翔鹤改二甲→翔鹤改二”
·启动时预先加载本地api_start2（用于离线查看各种图鉴等功能，实际游戏start2数据一直来自服务器）
·powerup, destoryitem2, destoryship修改为在正确接收服务器结果之后再修改数据集合
·调整自定义舰队编成的颜色在暗色主题下的异常

+++ ElectronicObserver/Observer/kcsapi/api_req_kousyou/destroyship.cs
@@ -9,34 +9,39 @@ namespace ElectronicObserver.Observer.kcsapi.api_req_kousyou {
 
 	public class destroyship : APIBase {
 
+		Dictionary<string, string> request;
 	
 		public override void OnRequestReceived( Dictionary<string, string> data ) {
-
-			KCDatabase db = KCDatabase.Instance;
-
-
-			//todo: ここに処理を書くのはみょんな感があるので、可能なら移動する
-
-			int shipID = int.Parse( data["api_ship_id"] );
-
-			db.Fleet.LoadFromRequest( APIName, data );
-
-			ShipData ship = db.Ships[shipID];
-			
-			Utility.Logger.Add( 2, ship.NameWithLevel + " 已解体。" );
-
-			for ( int i = 0; i < ship.Slot.Count; i++ ) {
-				if ( ship.Slot[i] != -1 )
-					db.Equipments.Remove( ship.Slot[i] );
-			}
-
-			db.Ships.Remove( shipID );
-
+			request = data;
 			
 			base.OnRequestReceived( data );
 		}
 
 		public override void OnResponseReceived( dynamic data ) {
+
+			if ( request != null )
+			{
+				KCDatabase db = KCDatabase.Instance;
+
+
+				//todo: ここに処理を書くのはみょんな感があるので、可能なら移動する
+
+				int shipID = int.Parse( request["api_ship_id"] );
+
+				db.Fleet.LoadFromRequest( APIName, request );
+
+				ShipData ship = db.Ships[shipID];
+
+				Utility.Logger.Add( 2, ship.NameWithLevel + " 已解体。" );
+
+				for ( int i = 0; i < ship.Slot.Count; i++ )
+				{
+					if ( ship.Slot[i] != -1 )
+						db.Equipments.Remove( ship.Slot[i] );
+				}
+
+				db.Ships.Remove( shipID );
+			}
 
 			KCDatabase.Instance.Material.LoadFromResponse( APIName, data.api_material );
